+++ src/index.js
@@ -27,6 +27,7 @@ import IFrame from "./iframe/iframe";
 import Table from "./table/table";
 import Video from "./video/video";
 import Travel from "./travel/travel";
+import VimeoPlayer from "./vimeo-player/vimeo-player";
 import "./index.css";
 
 export const renderScreen = (screen) => {
@@ -180,6 +181,16 @@ export const renderSlide = (slide) => {
     case "travel":
       return (
         <Travel
+          content={slide.content}
+          slide={slide}
+          run="1234"
+          slideDone={slideDone}
+          executionId="SLIDE_ID"
+        />
+      );
+    case "vimeo-player":
+      return (
+        <VimeoPlayer
           content={slide.content}
           slide={slide}
           run="1234"